Create README.md
The "Housing Price Prediction" project is a data analysis and machine learning endeavor focused on predicting housing prices in a specific region. In this case, we will use a dataset containing information about various housing properties in Bengaluru, India.

The main goal of this project is to build a reliable predictive model that can estimate housing prices based on different features. These features include the area (in square feet), the proximity to the main road, the availability of a basement, and the furnishing status of the property.